docs: decide not to use MapStruct for now (#10)
We (Eric and Tomas) has read up a bit on MapStruct and come to the
conclusion that it might become useful for us in the future. However, we
feel that it is wiser to postpone the decision until we know more about
the problem, i.e. when have some real world scenarios of data to map.

Using a tool like MapStruct will probably be most beneficial when there
are lots of fields to map and most of those fields are exactly the same
in both data structures. If there are lots of deviations there will
consequently be lots of annotations to configure the tooling to do the
right thing. If that proves to be the case for us, it is probably just
as simple to code the mapping by hand.

For now, it is better to keep it simple (KISS). With at bit of luck we
ain't gonna need it (YAGNI).

See: https://mapstruct.org/
See: https://en.wikipedia.org/wiki/KISS_principle
See: https://en.wikipedia.org/wiki/You_aren%27t_gonna_need_it

Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>
Co-authored-by: Tomas Robertsson <extern.tomas.robertsson@digg.se>