fix: use available version of PostgresSQL docker image (#19)
For some reason we can no longer pull `postgres:latest` from the Digg
Docker image registry. This makes the test suite fail when running on a
Digg computer. On GitHub everything seems to work, i.e. the test suite
passes. This is probably because we on GitHub go directly to the
official Docker registry and not via Digg.

To be more specific, when we run

    docker pull registry.digg.se:5050/postgres:latest

we get the following error:

    error pulling image configuration: download failed after attempts=1: unknown blob

This could potentially be a bug in Nexus. However, a workaround for now
is to use a version of the docker image that can actually be pulled from
Digg. One such version is `14-alpine` which was added to the Digg Nexus
server on 2025-02-17.

See: https://github.com/sonatype/nexus-public/issues/264
Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>