feat: stricter linting (#24)
This change set configures MegaLinter to be more strict. Specifically,
with the updated configuration we will fail the build in any of the
following situations:

 1. When there are any formatting issues detected
 2. When there are any automatic fixes suggested by any of the linters
 3. When Checkstyle finds any issues

Without these changes, formatting issues and automatic fixes are
detected and the source files are updated. However, the updates are not
committed - and worse, the issues are not even reported. Checkstyle
issues are reported as warnings but they are all too easy to miss and/or
ignore.

See: https://megalinter.io/latest/config-variables/
See: development/lint/google_checks.xml
See: https://megalinter.io/latest/descriptors/java_checkstyle/
Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>