feat: automate XML formatting (#21)
This change set adds XML formatting to our code quality checks via the
Prettier XML plugin. It is true that we already have XML linting via
MegaLinter, and in theory that XML linting tool supports formatting.
However, in practice it doesn't. I tried it out and it did not handle
long values in XML attributes very well. Also, in the official
documentation for that tool, it says that the formatting option is
broken.

Here is an excerpt from the xmllint documentation regarding the
formatting support:

> Especially in the absence of a DTD, this feature has never worked
> reliably and is fundamentally broken.

Instead of using a broken tool, here we have a solution built on
Prettier.

The XML formatting is implemented as a MegaLinter plugin using Prettier
and the Prettier XML plugin. By integrating with MegaLinter we ensure
that the formatting is run both by the code_quality.sh script and in
our GitHub pipelines. There is also support for running the formatting
from the command line and instructions for integrating with Visual
Studio Code.

See: https://gnome.pages.gitlab.gnome.org/libxml2/xmllint.html
See: https://megalinter.io/latest/plugins/
See: https://github.com/prettier/plugin-xml
Signed-off-by: Eric Thelin extern.eric.thelin@digg.se